# add nextjs optimized version