### Standard Idiomatic NextJS

- Time to Button Visible: 28 ms
- Time till counter incremented: 26128 ms
- Had to wait for network idle and click the button a second time: `TRUE`
- Second Click Response Time: 35 ms

### Optimized using Direct DOM API

- Time to Button Visible: 25 ms
- Time till counter incremented: 10 ms
- Had to wait for network idle and click the button a second time: `FALSE`
- Second Click Response Time: `N/A`